+++ TabletopTweaks-Base/NewContent/ContentAdder.cs
@@ -31,6 +31,7 @@ namespace TabletopTweaks.Base.NewContent {
                 Spells.ShadowEnchantment.AddShadowEnchantmentGreater();
                 Spells.MagicalTailSpells.AddNewMagicalTailSpells();
                 Spells.MagesDisjunction.AddMagesDisjunction();
+                Spells.StunningBarrierGreater.AddStunningBarrierGreater();
 
                 //Added early as some things depend on them for selections
                 Feats.MetamagicFeats.IntensifiedSpell.AddIntensifiedSpell();